% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{Colors}
\alias{Colors}
\alias{schneidr_purple}
\title{Functions for easily accessing preferred colors}
\usage{
schneidr_purple(alpha = 1)
}
\arguments{
\item{alpha}{A single number between 0 and 1}
}
\value{
A character vector with elements of 7 or 9 characters, "#" followed by the red, blue, green and optionally alpha values in hexadecimal.
}
\description{
Functions that return a color (or set of colors) of interest for use in data visualization.
}
\examples{

library(ggplot2)
mtcars[['gear']] <- factor(mtcars[['gear']])
ggplot(mtcars) +
geom_point(aes(x = wt, y = mpg),
           color = schneidr_purple(),
           size = 3) +
labs(title = "Fuel economy and weight",
     subtitle = "Data from the 1974 Motor Trend US magazine.",
     x = "Weight (1000 lbs)",
     y = "Fuel economy (mpg)")
}
