% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_save_plot.R
\name{quick_save_plot}
\alias{quick_save_plot}
\title{Save and preview a plot}
\usage{
quick_save_plot(
  plot = ggplot2::last_plot(),
  fname = tempfile(fileext = paste0(".", device)),
  inspect = TRUE,
  device = "png",
  dpi = 320,
  scale = 1,
  ...
)
}
\arguments{
\item{plot}{A plot object. If unspecified, uses the most recent printed plot.}

\item{fname}{A filename for the plot to be saved. Uses a temporary file in the user's temporary folder if unspecified.}

\item{inspect}{A logical defaulting to `TRUE` indicating whether to immediately open the file for inspection.}

\item{device}{Device to use. Can either be a device function (e.g. png()), or one of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only). Defaults to 'png'.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320), "print" (300), or "screen" (72). Applies only to raster output types.}

\item{scale}{Multiplicative scaling factor.}

\item{\dots}{Additional arguments passed on to the `ggsave` function.}
}
\value{
Saves a plot to a temporary file in the user's temp directory
}
\description{
Convenience function to quickly save a plot to an image file and immediately inspect it. Based on the function `ggpreview` by Tristan (TJ) Mahr.
}
\examples{
ggplot(mtcars, aes(mpg, wt)) + geom_point()
quick_save_plot()
}
