% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.R
\name{Derivatives}
\alias{Derivatives}
\alias{get_hessian}
\alias{get_jacobian}
\title{Tools for differentiating mathematical functions}
\usage{
get_hessian(f, as_matrix = FALSE, eval_at = NULL)

get_jacobian(f, as_matrix = FALSE, eval_at = NULL)
}
\arguments{
\item{f}{A mathematical expression generated with the \code{expression} function (e.g. \code{expression(y*sin(x))})}

\item{as_matrix}{Whether to return the symbolic matrix of derivatives as a list of expressions or as a character matrix.}

\item{eval_at}{A named list giving a point at which the matrix of derivatives is to be evaluated (e.g. \code{list(x = 1, y = 2)})}
}
\value{
If \code{as_matrix = TRUE}, a character matrix showing the symbolic matrix of derivatives. If \code{as_matrix = FALSE}, the result is a nested list of expressions. To retrieve the entry in row i column j of the matrix, select \code{hessian_result[[i]][[j]]}. If an evaluation point is specified with \code{eval_at}, the matrix of derivatives is evaluated and the result is a numeric matrix.
}
\description{
Obtain a function's Jacobian (matrix of first derivatives) or Hessian (matrix of second partial derivatives).
}
\examples{
# Create a mathematical expression
fn <- expression(a * log(p/(1-p)))

# Get the symbolic Hessian
## As a nested list
  get_hessian(fn)

## As a character matrix
  get_hessian(fn, as_matrix = TRUE)

# Evaluated the Hessian at a specific point

  get_hessian(fn, eval_at = list(a = 100, p = 0.5))

# Get the symbolic Jacobian
## As a nested list
   get_jacobian(fn)

## As a character matrix
   get_jacobian(fn, as_matrix = TRUE)

# Evaluate the Jacobian at a specific point

   get_jacobian(fn, eval_at = list(a = 100, p = 0.5))
}
\concept{math tools}
